+++ CHANGELOG.md
@@ -25,6 +25,9 @@
 * Fixed multithreading JS output for targets `bundler`, `deno` and `module`.
   [#4685](https://github.com/wasm-bindgen/wasm-bindgen/pull/4685)
 
+* Fixed `TextDe/Encoder` detection for audio worklet use-cases.
+  [#4703](https://github.com/wasm-bindgen/wasm-bindgen/pull/4703)
+
 * Fixed post-processing failures in case Std has debug assertions enabled.
   [#4705](https://github.com/wasm-bindgen/wasm-bindgen/pull/4705)